rename flexTable module and path to files, replace all standard functions to arrow fouctions
