+++ frontend/blocks/flex_table/flex_table.css
