+++ npm-debug.log
@@ -1,0 +1,46 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/home/user/.nvm/versions/node/v6.5.0/bin/node',
+1 verbose cli   '/home/user/.nvm/versions/node/v6.5.0/bin/npm',
+1 verbose cli   'run',
+1 verbose cli   'precommit' ]
+2 info using npm@3.10.7
+3 info using node@v6.5.0
+4 verbose run-script [ 'precommit' ]
+5 info lifecycle smartforms@1.0.0~precommit: smartforms@1.0.0
+6 verbose lifecycle smartforms@1.0.0~precommit: unsafe-perm in lifecycle true
+7 verbose lifecycle smartforms@1.0.0~precommit: PATH: /home/user/.nvm/versions/node/v6.5.0/lib/node_modules/npm/bin/node-gyp-bin:/home/user/Development/smartforms/node_modules/.bin:/home/user/.nvm/versions/node/v6.5.0/bin:/usr/lib/git-core:/usr/local/share/npm/bin:/home/user/.rvm/gems/ruby-2.1.2/bin:/home/user/.rvm/gems/ruby-2.1.2@global/bin:/home/user/.rvm/rubies/ruby-2.1.2/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/home/user/.rvm/bin
+8 verbose lifecycle smartforms@1.0.0~precommit: CWD: /home/user/Development/smartforms
+9 silly lifecycle smartforms@1.0.0~precommit: Args: [ '-c', 'npm run lint' ]
+10 silly lifecycle smartforms@1.0.0~precommit: Returned: code: 1  signal: null
+11 info lifecycle smartforms@1.0.0~precommit: Failed to exec precommit script
+12 verbose stack Error: smartforms@1.0.0 precommit: `npm run lint`
+12 verbose stack Exit status 1
+12 verbose stack     at EventEmitter.<anonymous> (/home/user/.nvm/versions/node/v6.5.0/lib/node_modules/npm/lib/utils/lifecycle.js:255:16)
+12 verbose stack     at emitTwo (events.js:106:13)
+12 verbose stack     at EventEmitter.emit (events.js:191:7)
+12 verbose stack     at ChildProcess.<anonymous> (/home/user/.nvm/versions/node/v6.5.0/lib/node_modules/npm/lib/utils/spawn.js:40:14)
+12 verbose stack     at emitTwo (events.js:106:13)
+12 verbose stack     at ChildProcess.emit (events.js:191:7)
+12 verbose stack     at maybeClose (internal/child_process.js:877:16)
+12 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
+13 verbose pkgid smartforms@1.0.0
+14 verbose cwd /home/user/Development/smartforms
+15 error Linux 4.4.0-36-generic
+16 error argv "/home/user/.nvm/versions/node/v6.5.0/bin/node" "/home/user/.nvm/versions/node/v6.5.0/bin/npm" "run" "precommit"
+17 error node v6.5.0
+18 error npm  v3.10.7
+19 error code ELIFECYCLE
+20 error smartforms@1.0.0 precommit: `npm run lint`
+20 error Exit status 1
+21 error Failed at the smartforms@1.0.0 precommit script 'npm run lint'.
+21 error Make sure you have the latest version of node.js and npm installed.
+21 error If you do, this is most likely a problem with the smartforms package,
+21 error not with npm itself.
+21 error Tell the author that this fails on your system:
+21 error     npm run lint
+21 error You can get information on how to open an issue for this project with:
+21 error     npm bugs smartforms
+21 error Or if that isn't available, you can get their info via:
+21 error     npm owner ls smartforms
+21 error There is likely additional logging output above.
+22 verbose exit [ 1, true ]
